feat : add new usecase for home modules and make modification to several files

+++ lib/features/home/presentation/widgets/dashboard_widget.dart
@@ -9,7 +9,7 @@ import '../../../../core/routes/app_routes.dart';
 
 class DashboardWidget extends StatelessWidget {
   // <--- Public Class
-  final int totalPendapatan;
+  final double totalPendapatan;
   final int totalTransaksi;
   final int motorCount;
   final int mobilCount;
